+++ embedded/benchmark/src/main.rs
@@ -4,6 +4,7 @@
 //! Benchmarks for `woxml::XmlWriter`
 
 extern crate alloc;
+extern crate bytes;
 
 use alloc::vec::Vec;
 use alloc::vec;
@@ -13,6 +14,31 @@ use woxml::XmlWriter;
 
 fn create_xml(
     writer: &mut XmlWriter<'_, Vec<u8>>,
+    nsmap: &Vec<(Option<&'static str>, &'static str)>,
+) {
+    _ = writer.begin_elem("OTDS");
+    _ = writer.ns_decl(nsmap);
+    _ = writer.comment("have a nice day");
+    writer.set_namespace("st");
+    _ = writer.empty_elem("success");
+    _ = writer.begin_elem("node");
+    _ = writer.attr_esc("name", "\"123\"");
+    _ = writer.attr("id", "abc");
+    _ = writer.attr("'unescaped'", "\"123\""); // this WILL generate invalid xml
+    _ = writer.text("'text'");
+    _ = writer.end_elem();
+    writer.unset_namespace();
+    _ = writer.comment("comment");
+    _ = writer.begin_elem("stuff");
+    _ = writer.cdata("blablab");
+    _ = writer.end_elem();
+    _ = writer.end_elem();
+    _ = writer.close();
+    _ = writer.flush();
+}
+
+fn create_xml_bytes(
+    writer: &mut XmlWriter<'_, bytes::BytesMut>,
     nsmap: &Vec<(Option<&'static str>, &'static str)>,
 ) {
     _ = writer.begin_elem("OTDS");
@@ -49,7 +75,20 @@ fn main() {
         create_xml(&mut xml, &nsmap);
     }) {
         Ok(ticks) => {
-            info!("compact mode took {} per iteration", ticks);
+            info!("compact with Vec took {} per iteration", ticks);
+        }
+        Err(err) => {
+            error!("benchmark returned: {}", err);
+            exit(ExitCode::FAILURE);
+        }
+    }
+
+    match ariel_os::bench::benchmark(100, || {
+        let mut xml = XmlWriter::compact_mode(bytes::BytesMut::new());
+        create_xml_bytes(&mut xml, &nsmap);
+    }) {
+        Ok(ticks) => {
+            info!("compact with BytesMut took {} per iteration", ticks);
         }
         Err(err) => {
             error!("benchmark returned: {}", err);
@@ -62,7 +101,20 @@ fn main() {
         create_xml(&mut xml, &nsmap);
     }) {
         Ok(ticks) => {
-            info!("pretty mode took {} per iteration", ticks);
+            info!("pretty with Vec took {} per iteration", ticks);
+        }
+        Err(err) => {
+            error!("benchmark returned: {}", err);
+            exit(ExitCode::FAILURE);
+        }
+    }
+
+    match ariel_os::bench::benchmark(100, || {
+        let mut xml = XmlWriter::pretty_mode(bytes::BytesMut::new());
+        create_xml_bytes(&mut xml, &nsmap);
+    }) {
+        Ok(ticks) => {
+            info!("compact with BytesMut took {} per iteration", ticks);
         }
         Err(err) => {
             error!("benchmark returned: {}", err);